Unify computation of history key
